+++ internal/usecase/transaction/update/update.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 )
 
-func (u *Usecase) Update(ctx context.Context, id int, amount float64) error {
+func (u *Usecase) Update(ctx context.Context, id int, amount int64) error {
 	if err := u.transactionRepo.UpdateTransaction(ctx, id, amount); err != nil {
 		return fmt.Errorf("Failed to edit transaction: %w", err)
 	}
